extract coldefs and mock member data to utils files, decrease line height in overflow text
