+++ client/src/utils/mockMemberData.js
@@ -1,0 +1,56 @@
+export const mockMemberData = [
+  {
+    name: 'Amitahwieufhfawileuhfaiweuhf',
+    email: 'sawhney4@illinois.edu',
+    phone: '(309) 838-5466',
+    netid: 'sawhney4',
+    uin: 659426199,
+    major: 'CS + Math',
+    birthdate: '02/14/2002',
+    github: 'amit-sawhneyawefaiwehfiauwehfiuaehfiuawhegoiawuehoiawgueh',
+    snapchat: 'amit_sawhney02',
+    instagram: 'amit__02',
+    grad: 'Spring 2024 (Freshman)',
+    generation: 'Fall 2020',
+    location: 'On-campus',
+    role: 'Software Developer',
+    level: 'Member',
+    status: 'Active',
+  },
+  {
+    name: 'Alice',
+    email: 'alicesf2@illinois.edu',
+    phone: '(123)-456-7891',
+    netid: 'alicesf2',
+    uin: 123456789,
+    major: 'CS',
+    birthdate: '01/02/3000',
+    github: 'alicesf2',
+    snapchat: 'alicesf2',
+    instagram: 'alicesf2',
+    grad: 'Fall 2020 (Alum)',
+    generation: 'Fall 2018',
+    location: 'On-campus',
+    role: 'Tech Lead',
+    level: 'Member',
+    status: 'Active',
+  },
+  {
+    name: 'Jefffrey',
+    email: 'alicesf2@illinois.edu',
+    phone: '(123)-456-7891',
+    netid: 'alicesf2',
+    uin: 123456789,
+    major: 'CS',
+    birthdate: '01/02/1999',
+    github: 'alicesf2',
+    snapchat: 'alicesf2',
+    instagram: 'alicesf2',
+    grad: 'Spring 2020',
+    generation: 'Spring 2024',
+    location: 'Remote',
+    role: 'Tech Lead',
+    level: 'Member',
+    status: 'Hiatus',
+  },
+];
